+++ custom_components/climate_group_helper/const.py
@@ -200,14 +200,14 @@ ATTR_SCHEDULE_ENTITY = "schedule_entity"
 # Extra attribute keys
 ATTR_ACTIVE_SCHEDULE_ENTITY = "active_schedule_entity"
 ATTR_ASSUMED_STATE = "assumed_state"
-ATTR_BLOCKING_REASON = "blocking_reason"
+ATTR_BLOCKING_SOURCES = "blocking_sources"
 ATTR_CURRENT_HVAC_MODES = "current_hvac_modes"
 ATTR_ISOLATED_MEMBERS = "isolated_members"
 ATTR_LAST_ACTIVE_HVAC_MODE = "last_active_hvac_mode"
 ATTR_OOB_MEMBERS = "oob_members"
 ATTR_ACTIVE_OVERRIDE = "active_override"
 ATTR_MASTER_FALLBACK_ACTIVE = "master_fallback_active"
-ATTR_SCHEDULE_OVERRIDE_ACTIVE = "schedule_override_active"
+ATTR_ACTIVE_OVERRIDE_END = "active_override_end"
 
 # Attribute to service call mapping
 ATTR_SERVICE_MAP = {